+++ packages/language-services/src/features/__tests__/do-complete-modules.test.ts
@@ -59,6 +59,92 @@ test("suggests built-in sass modules", async () => {
 	);
 });
 
+test("suggest sass built-ins that are forwarded by the stylesheet that is @used", async () => {
+	const one = fileSystemProvider.createDocument(['@forward "sass:math";'], {
+		uri: "test.scss",
+	});
+	const two = fileSystemProvider.createDocument([
+		'@use "./test";',
+		"$var: test.",
+	]);
+
+	// emulate scanner of language service which adds workspace documents to the cache
+	ls.parseStylesheet(one);
+	ls.parseStylesheet(two);
+
+	const { items } = await ls.doComplete(two, Position.create(1, 11));
+	assert.notEqual(
+		items.length,
+		0,
+		"Expected to get completions from the sass:math module",
+	);
+
+	// Quick sampling of the results
+	assert.deepStrictEqual(
+		items.find((annotation) => annotation.label === "$pi"),
+		{
+			documentation: {
+				kind: "markdown",
+				value:
+					"The value of the mathematical constant **π**.\n\n[Sass documentation](https://sass-lang.com/documentation/modules/math#$pi)",
+			},
+			filterText: "test.$pi",
+			insertText: ".$pi",
+			insertTextFormat: InsertTextFormat.PlainText,
+			kind: CompletionItemKind.Variable,
+			label: "$pi",
+			labelDetails: {
+				detail: undefined,
+			},
+		},
+	);
+});
+
+test("suggest sass built-ins that are forwarded with a prefix", async () => {
+	const one = fileSystemProvider.createDocument(
+		['@forward "sass:math" as math-*;'],
+		{
+			uri: "test.scss",
+		},
+	);
+	const two = fileSystemProvider.createDocument([
+		'@use "./test";',
+		"$var: test.",
+	]);
+
+	// emulate scanner of language service which adds workspace documents to the cache
+	ls.parseStylesheet(one);
+	ls.parseStylesheet(two);
+
+	const { items } = await ls.doComplete(two, Position.create(1, 11));
+	assert.notEqual(
+		items.length,
+		0,
+		"Expected to get completions from the sass:math module",
+	);
+
+	// Quick sampling of the results
+	assert.deepStrictEqual(
+		items.find((annotation) => annotation.label === "$math-pi"),
+		{
+			documentation: {
+				kind: "markdown",
+				value:
+					"The value of the mathematical constant **π**.\n\n[Sass documentation](https://sass-lang.com/documentation/modules/math#$pi)",
+			},
+			filterText: "test.$math-pi",
+			insertText: ".$math-pi",
+			insertTextFormat: InsertTextFormat.PlainText,
+			kind: CompletionItemKind.Variable,
+			label: "$math-pi",
+			labelDetails: {
+				detail: undefined,
+			},
+		},
+	);
+	assert.ok(items.find((a) => a.label === "math-clamp"));
+});
+
 test("should suggest symbol from a different document via @use", async () => {
 	const one = fileSystemProvider.createDocument("$primary: limegreen;", {
 		uri: "one.scss",
@@ -1090,4 +1176,29 @@ test("should suggest symbol from a different document via @use with wildcard ali
 			tags: [],
 		},
 	);
+});
+
+test("does not suggest sass globals if suggestFromUseOnly is true", async () => {
+	ls.configure({
+		completionSettings: {
+			suggestFromUseOnly: true,
+		},
+	});
+
+	const document = fileSystemProvider.createDocument("@debug co");
+	const { items } = await ls.doComplete(document, Position.create(0, 9));
+	assert.isUndefined(items.find((item) => item.label === "comparable"));
+});
+
+// We don't call the upstream for suggestions here since we got complaints about duplicates
+test.skip("does suggest sass globals if suggestFromUseOnly is false", async () => {
+	ls.configure({
+		completionSettings: {
+			suggestFromUseOnly: false,
+		},
+	});
+
+	const document = fileSystemProvider.createDocument("@debug co");
+	const { items } = await ls.doComplete(document, Position.create(0, 9));
+	assert.isOk(items.find((item) => item.label === "comparable"));
 });