+++ packages/language-services/src/language-services-types.ts
@@ -145,7 +145,7 @@ export interface LanguageService {
 		document: TextDocument,
 		positions: Position[],
 	): Promise<SelectionRange[]>;
-	hasCached(uri: URI): boolean;
+	getCachedTextDocument(uri: URI): TextDocument | undefined;
 	/**
 	 * Utility function to reparse an updated document.
 	 * Like {@link LanguageService.parseStylesheet}, but returns nothing.