+++ packages/language-services/CHANGELOG.md
@@ -1,3 +1,57 @@
+## 1.3.4 (2024-08-24)
+
+
+### 🩹 Fixes
+
+- race condition in workspace scanner
+
+- exclude sass globals if suggestFromUseOnly
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/vscode-css-languageservice to 1.3.1
+
+## 1.3.3 (2024-08-20)
+
+
+### 🩹 Fixes
+
+- a race condition that caused partials not to be parsed sometimes
+
+## 1.3.2 (2024-08-20)
+
+
+### 🩹 Fixes
+
+- apply completion settings
+
+## 1.3.1 (2024-08-11)
+
+
+### 🩹 Fixes
+
+- adjust when suggesting functions, variables in control flow
+
+- add prefix support to forwarded sass built-ins
+
+## 1.3.0 (2024-08-07)
+
+
+### 🚀 Features
+
+- feature update for language server
+
+
+### 🩹 Fixes
+
+- suggest from forwarded sass built-ins
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/vscode-css-languageservice to 1.3.0
+
 ## 1.2.0 (2024-08-04)