+++ packages/vscode-css-languageservice/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.3.1 (2024-08-24)
+
+
+### 🩹 Fixes
+
+- exclude sass globals if suggestFromUseOnly
+
+## 1.3.0 (2024-08-07)
+
+
+### 🚀 Features
+
+- feature update for language server
+
 ## 1.2.0 (2024-08-04)